+++ client/src/util/misc.ts
@@ -271,49 +271,10 @@ export function GetGradeFromPercent<I extends IDStrings = IDStrings>(
 	throw new Error(`Could not resolve grade for percent ${percent} on game ${game}`);
 }
 
-/**
- * Calculates the PikaGreatFunction, used in BPI. I have no idea what this does.
- */
-function BPIPikaGreatFn(score: integer, max: integer) {
-	return score === max ? max * 0.8 : 1 + (score / max - 0.5) / (1 - score / max);
-}
-
-export function CalculateBPI(
-	kaidenEx: integer,
-	wrEx: integer,
-	yourEx: integer,
-	max: integer,
-	pc: number | null
-) {
-	let powCoef = pc ?? 1.175;
-	if (powCoef === -1) {
-		powCoef = 1.175;
+export function ConditionalLeadingSpace(maybeStr: string | null) {
+	if (maybeStr === null) {
+		return "";
 	}
 
-	const yourPGF = BPIPikaGreatFn(yourEx, max);
-	const kaidenPGF = BPIPikaGreatFn(kaidenEx, max);
-	const wrPGF = BPIPikaGreatFn(wrEx, max);
-
-	// no idea what these var names are
-	const _s_ = yourPGF / kaidenPGF;
-	const _z_ = wrPGF / kaidenPGF;
-
-	const isBetterThanKavg = yourEx >= kaidenEx;
-
-	// this line of code isn't mine, and that's why it's *really* bad here.
-	const bpi =
-		Math.round(
-			(isBetterThanKavg ? 100 : -100) *
-				Math.pow(
-					(isBetterThanKavg ? Math.log(_s_) : -Math.log(_s_)) / Math.log(_z_),
-					powCoef
-				) *
-				100
-		) / 100;
-
-	if (bpi < -15) {
-		return -15;
-	}
-
-	return bpi;
+	return ` ${maybeStr}`;
 }